Update dependencies and fix plugin integration tests
